Remove definer concept; what's left, move to "handler" package
Also deleted MockRouter and redundant tests.
