+++ src/main/java/HttpServer/cobspec/Server.java
@@ -1,7 +1,7 @@
 package HttpServer.cobspec;
 
 import HttpServer.core.SynchronousListener;
-import HttpServer.core.definer.*;
+import HttpServer.core.handler.*;
 import HttpServer.core.router.Router;
 import HttpServer.core.utility.CommandLineParser;
 import HttpServer.core.utility.Logger;
@@ -18,43 +18,48 @@ public class Server {
         CommandLineParser commandLineParser = new CommandLineParser(args, port, directory, logger).invoke();
         port = commandLineParser.getPort();
         directory = commandLineParser.getDirectory();
-
-        IRouteDefiner definer = new IRouteDefiner() {
-            private final Router router;
-
-            {
-                Router blankRouter = new Router();
-                FileRouteDefiner fileDefiner = new FileRouteDefiner("./cob_spec/public", blankRouter);
-                TeaRouteDefiner teaDefiner = new TeaRouteDefiner(fileDefiner.getRouter());
-                HashMap<String, String> redirectionMap = new HashMap<String, String>();
-                redirectionMap.put("/redirect", "/");
-                RedirectDefiner redirectDefiner = new RedirectDefiner(teaDefiner.getRouter(), redirectionMap);
-                this.router = redirectDefiner.getRouter();
-
-                // Options tests are a concern of the router, in my opinion. Not set with a definer.
-                Handler okHandler = new FunctionalHandler(200);
-                router.defineRoute("/method_options", "GET", okHandler);
-                router.defineRoute("/method_options", "HEAD", okHandler);
-                router.defineRoute("/method_options", "POST", okHandler);
-                router.defineRoute("/method_options", "OPTIONS", okHandler);
-                router.defineRoute("/method_options", "PUT", okHandler);
-                router.defineRoute("/method_options2", "GET", okHandler);
-                router.defineRoute("/method_options2", "OPTIONS", okHandler);
-                router.defineRoute("/", "GET", new FunctionalHandler(200));
-                router.defineRoute("/", "HEAD", new FunctionalHandler(200));
-                router.defineRoute("/form", "PUT", new FunctionalHandler(200));
-            }
-
-            @Override
-            public Router getRouter() {
-                return router;
-            }
-
-        };
+        Router router = new Router();
+        defineRoutes(router);
 
         // Listening. Opens a server socket, accepts a connection, reads/writes, closes the connection. Repeat.
-        SynchronousListener listener = new SynchronousListener(port, directory, definer, logger);
+        SynchronousListener listener = new SynchronousListener(port, directory, router, logger);
         listener.start();
     }
 
+    private static void defineRoutes(Router router) {
+        // Routes for satisfying cob_spec acceptance test suite.
+
+        // SimpleGet, SimpleHead
+        Handler okHandler = new FunctionalHandler(200);
+        router.defineRoute("/", "GET", okHandler);
+        router.defineRoute("/", "HEAD", okHandler);
+
+        // SimplePut, SimplePost
+        router.defineRoute("/form", "PUT", okHandler);
+        router.defineRoute("/form", "POST", okHandler);
+
+        // FourEightTeen
+        Handler teaHandler = new TeaHandler();
+        router.defineRoute("/tea", "GET", teaHandler);
+        router.defineRoute("/coffee", "GET", teaHandler);
+
+        // RedirectPath
+        HashMap<String, String> redirectionMap = new HashMap<>();
+        redirectionMap.put("/redirect", "/");
+        Handler redirectHandler = new RedirectHandler(redirectionMap);
+        router.defineRoute("/redirect", "GET", redirectHandler);
+
+        // SimpleOption
+        router.defineRoute("/method_options", "GET", okHandler);
+        router.defineRoute("/method_options", "HEAD", okHandler);
+        router.defineRoute("/method_options", "POST", okHandler);
+        router.defineRoute("/method_options", "OPTIONS", okHandler);
+        router.defineRoute("/method_options", "PUT", okHandler);
+        router.defineRoute("/method_options2", "GET", okHandler);
+        router.defineRoute("/method_options2", "OPTIONS", okHandler);
+
+        // MethodNotAllowed
+        router.defineRoute("/file1", "GET", okHandler);
+        router.defineRoute("/text-file.txt", "GET", okHandler);
+    }
 }
