+++ src/main/java/HttpServer/core/handler/DirectoryHandler.java
@@ -1,4 +1,4 @@
-package HttpServer.core.definer;
+package HttpServer.core.handler;
 
 import HttpServer.core.request.Request;
 import HttpServer.core.response.Response;
