+++ src/test/java/HttpServer/core/handler/RedirectHandler.java
@@ -20,8 +20,8 @@
 //        HashMap<String, String> redirectionMap = new HashMap<>();
 //        redirectionMap.put("/redirect", "/");
 //        Router blankRouter = new Router();
-//        RedirectDefiner definer = new RedirectDefiner(blankRouter, redirectionMap);
-//        this.router = definer.getRouter();
+//        RedirectDefiner handler = new RedirectDefiner(blankRouter, redirectionMap);
+//        this.router = handler.getRouter();
 //
 //        this.request = new Request();
 //        request.setUri("/redirect");
@@ -39,5 +39,24 @@
 //        Response response = router.route(request);
 //        assertEquals("/", response.getHeader("Location"));
 //    }
+
+
+//import org.junit.Test;
+//
+//@Test
+//public void itWritesARedirectStatusLine() throws IOException {
+//        String requestLine = "GET /redirect HTTP/1.1";
+//        MockClient client = (MockClient) runner.apply(requestLine);
+//        String expectedStatusLine = "HTTP/1.1 302 Found";
+//        assertEquals(expectedStatusLine, client.output.get(0));
+//        }
+//
+//@Test
+//public void itWritesARedirectLocation() throws IOException {
+//        String requestLine = "GET /redirect HTTP/1.1";
+//        MockClient client = (MockClient) runner.apply(requestLine);
+//        String expectedLocationLine = "Location: /";
+//        assertEquals(expectedLocationLine, client.output.get(1));
+//        }
 //
 //}
