+++ src/test/java/HttpServer/core/handler/TeaHandlerTest.java
@@ -1,4 +1,4 @@
-package HttpServer.core.definer;
+package HttpServer.core.handler;
 
 import HttpServer.core.request.Request;
 import HttpServer.core.response.Response;
